优化了后端创建剧本库文件时候的时候的id为18位整数型数字

+++ src/app/models/script.py
@@ -22,7 +22,7 @@ class ScriptFile(Base):
     """剧本库中的具体文件"""
     __tablename__ = "script_files"
 
-    id = Column(Integer, primary_key=True, index=True)
+    id = Column(BigInteger, primary_key=True, index=True)
     library_id = Column(BigInteger, ForeignKey("script_libraries.id"), nullable=False)
     filename = Column(String, nullable=False)
     file_url = Column(String, nullable=True)  # Minio 访问链接
